Only store the first 80 characters of the commit
This is a pretty standard practice, to take either the first line (or if
it's a long first line, the first 80 characters) and use that to
identify the commit without having to store the whole body. (We have the
SHA so we can always point them to the URL if they want to see the
rest).